Configure user db backend.

+++ user/user-db/src/main/java/liquid/user/db/repository/GroupMemberRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by Tao Ma on 3/18/15.
  */
-public interface GroupMemberRepository extends CrudRepository<GroupMember, Long> {
+public interface GroupMemberRepository extends CrudRepository<GroupMember, Integer> {
 }
