+++ user/user-db/src/main/java/liquid/user/db/repository/GroupRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by Tao Ma on 3/18/15.
  */
-public interface GroupRepository extends CrudRepository<Group, Long> {
+public interface GroupRepository extends CrudRepository<Group, Integer> {
 }
